+++ app/src/main/java/de/baumann/browser/activity/BrowserActivity.java
@@ -1215,7 +1215,7 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
                 assert url != null;
                 if (url.startsWith("data:")) {
                     DataURIParser dataURIParser = new DataURIParser(url);
-                    HelperUnit.saveDataURI(activity, title, url, dataURIParser, dialog_overflow);
+                    HelperUnit.saveDataURI(activity, dataURIParser, dialog_overflow);
                 } else HelperUnit.saveAs(activity, url, null, dialog_overflow, ninjaWebView);
             }
         });
@@ -1589,24 +1589,6 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
                     builderSubMenu = new MaterialAlertDialogBuilder(context);
 
                     View dialogViewSubMenu = View.inflate(context, R.layout.dialog_edit, null);
-
-                    LinearLayout textGroupEdit = dialogViewSubMenu.findViewById(R.id.textGroupEdit);
-                    TextView menuURLEdit = dialogViewSubMenu.findViewById(R.id.menuURLEdit);
-                    menuURLEdit.setText(url);
-                    menuURLEdit.setEllipsize(TextUtils.TruncateAt.MARQUEE);
-                    menuURLEdit.setSingleLine(true);
-                    menuURLEdit.setMarqueeRepeatLimit(1);
-                    menuURLEdit.setSelected(true);
-                    textGroupEdit.setOnClickListener(v -> {
-                        menuURLEdit.setEllipsize(TextUtils.TruncateAt.MARQUEE);
-                        menuURLEdit.setSingleLine(true);
-                        menuURLEdit.setMarqueeRepeatLimit(1);
-                        menuURLEdit.setSelected(true);
-                    });
-                    TextView menuTitleEdit = dialogViewSubMenu.findViewById(R.id.menuTitleEdit);
-                    menuTitleEdit.setText(HelperUnit.domain(url));
-                    FaviconHelper.setFavicon(context, dialogViewSubMenu, null, R.id.menu_icon, R.drawable.icon_edit);
-
                     LinearLayout editButtonsLayout = dialogViewSubMenu.findViewById(R.id.editButtonsLayout);
                     editButtonsLayout.setVisibility(View.VISIBLE);
                     TextInputLayout editTopLayout = dialogViewSubMenu.findViewById(R.id.editTopLayout);
@@ -1667,6 +1649,8 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
                     newIcon = recordList.get(location).getIconColor();
                     HelperUnit.setFilterIcons(context, ib_icon, newIcon);
 
+                    builderSubMenu.setTitle(R.string.menu_edit);
+                    builderSubMenu.setIcon(R.drawable.icon_edit);
                     builderSubMenu.setView(dialogViewSubMenu);
                     dialogSubMenu = builderSubMenu.create();
                     dialogSubMenu.show();
@@ -2544,9 +2528,6 @@ public class BrowserActivity extends AppCompatActivity implements BrowserControl
             GridItem item_04 = new GridItem( context.getString(R.string.dialog_translate), R.drawable.icon_translate);
 
             View dialogView = View.inflate(context, R.layout.dialog_menu, null);
-
-            MaterialCardView albumCardView = dialogView.findViewById(R.id.albumCardView);
-            albumCardView.setVisibility(View.GONE);
 
             MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(context);
             builder.setIcon(R.drawable.icon_share);