+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v 15 ... (WIP)
 - new: auto deny cookie-banners (@https://github.com/woheller69/browser)
 - new: sort startsite and history by domain
+- fix: layout and colors
 - fix: Error downloading blob-files (@https://github.com/woheller69/browser)
 - fix: Downloading files that requires Referer header #950